backend/parla/requirements.txt
botocore
django-admin-interface
django-dynamic-admin-forms
django-nested-admin
django-jalali-date 
jdatetime 
requests
djoser
django-storages
boto3
django-ckeditor-5
django_admin_contexts
pillow
django
djangorestframework
python-decouple
django-jalali-date
psycopg2-binary
django-admin-interface
django-dynamic-admin-forms
django-nested-admin
django-admin-contexts
djangorestframework-simplejwt
python-decouple
